admin-service/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9

# Authentication
python-jose>=3.3.0
python-multipart>=0.0.6

# HTTP Client
httpx>=0.25.1

# Miscellaneous
starlette>=0.27.0
python-dotenv>=1.0.0